Add 'get_stats' method to BaseItem and subclasses
An abstract 'get_stats' method was added to the BaseItem in 'utils.py', which is meant to be overridden by its subclasses in 'items.py'. This new method produces a string representation of individual item stats. It will be utilized in 'game.py' when generating a list of character's items in a human-readable format.