Add @property decorators to character and item methods
The @property decorators were added to some methods in the characters.py and items.py files. This change allows class attributes like 'name', 'special_attack_name', 'passive_name' and 'is_unique_passive' to be treated as read-only properties. This ensures safer and controlled access to these attributes to retain the integrity of the classes.